Zeroconf communication and encryption
Code refactored.
Reverted changes to packaging.
