+++ src/main/java/com/shuffle/scplayer/SCPlayerMain.java
@@ -12,6 +12,9 @@ import java.util.UUID;
 import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.Mixer;
 
+import com.shuffle.scplayer.core.zeroconf.SpotifyZeroConfProviderFactory;
+import com.shuffle.scplayer.core.zeroconf.ZeroConfService;
+import com.shuffle.scplayer.core.zeroconf.ZeroConfServiceAvahiBin;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.log4j.Level;
@@ -33,7 +36,7 @@ import com.shuffle.scplayer.core.SpotifyConnectPlayerImpl;
 import com.shuffle.scplayer.core.Track;
 import com.shuffle.scplayer.jna.SpotifyLibrary.SpBitrate;
 import com.shuffle.scplayer.web.PlayerWebServerIntegration;
-import com.shuffle.scplayer.core.SpotifyConnectZeroConf;
+import com.shuffle.scplayer.core.zeroconf.SpotifyZeroConfServer;
 
 public class SCPlayerMain {
 	private static final transient Log log = LogFactory.getLog(SCPlayerMain.class);
@@ -46,6 +49,7 @@ public class SCPlayerMain {
 	private static String blob;
 	private static String playerName;
 	private static String deviceId = UUID.randomUUID().toString();
+
 	private static boolean rememberMe;
 	private static JsonObject credentialsJson; 
 	static {
@@ -242,7 +246,10 @@ public class SCPlayerMain {
 		player.setPlayerName(playerName);
 
 		// starting zeroconf service
-		SpotifyConnectZeroConf zeroConf = new SpotifyConnectZeroConf(player);
+		String zeroConfImpl = System.getProperty("zeroconf","library");
+		ZeroConfService zeroConfService = new ZeroConfServiceAvahiBin();
+		SpotifyZeroConfProviderFactory providerFactory = new SpotifyZeroConfProviderFactory();
+		SpotifyZeroConfServer zeroConf = new SpotifyZeroConfServer(providerFactory.create(zeroConfImpl, player), zeroConfService);
 		zeroConf.runServer();
 
 		String mixerString = System.getProperty("mixer", "0");
