+++ src/main/java/com/shuffle/scplayer/core/zeroconf/SpotifyZeroConfProvider.java
@@ -1,0 +1,6 @@
+package com.shuffle.scplayer.core.zeroconf;
+
+public interface SpotifyZeroConfProvider {
+    SpotifyZeroConfVars getVars();
+    void loginZeroConf(String username, String blob, String clientKey) throws Exception;
+}
