+++ src/main/java/com/shuffle/scplayer/core/zeroconf/SpotifyZeroConfVars.java
@@ -1,0 +1,44 @@
+package com.shuffle.scplayer.core.zeroconf;
+
+public class SpotifyZeroConfVars {
+    private String publicKey;
+    private String deviceId;
+    private String activeUser;
+    private String remoteName;
+    private String accountReq;
+    private int deviceType;
+
+    public SpotifyZeroConfVars(String publicKey, String deviceId, String activeUser, String remoteName, String accountReq, int deviceType) {
+        this.publicKey = publicKey;
+        this.deviceId = deviceId;
+        this.activeUser = activeUser;
+        this.remoteName = remoteName;
+        this.accountReq = accountReq;
+        this.deviceType = deviceType;
+    }
+
+    public String getPublicKey() {
+        return publicKey;
+    }
+
+    public String getDeviceId() {
+        return deviceId;
+    }
+
+    public String getActiveUser() {
+        return activeUser;
+    }
+
+    public String getRemoteName() {
+        return remoteName;
+    }
+
+    public String getAccountReq() {
+        return accountReq;
+    }
+
+    public int getDeviceType() {
+        return deviceType;
+    }
+
+}
