+++ src/main/java/com/shuffle/scplayer/core/zeroconf/ZeroConfService.java
@@ -1,0 +1,5 @@
+package com.shuffle.scplayer.core.zeroconf;
+
+public interface ZeroConfService {
+    void publishService(String name, String type, int port, String text);
+}
